refactor: extract media upload operations into lib/client/media.ts
